+++ shell/run.sh
@@ -1,8 +1,25 @@
 #!/usr/bin/env bash
 
-source ./pre_run.sh
+WORKING_DIR="src"
 
-printf "\nStarting app\n\n"
+cd .. && source .venv/bin/activate || exit 1
 
-python3 aideas/main.py
-#python3 aideas/main.py --agents translation
+printf "\nExporting environment\n"
+
+set -a
+source .env
+set +a
+
+export PYTHONUNBUFFERED=1
+
+cd "$WORKING_DIR" || (printf "\nCould not change to working dir: %s\n" "$WORKING_DIR" && exit 1)
+
+printf "\nWorking from: %s\n" "$(pwd)"
+
+if [ "$WEB_APP" = true ] || [ "$WEB_APP" = "true" ] ; then
+  printf "\nStarting web app\n\n"
+  python3 aideas/web.py  # --agents translation
+else
+  printf "\nStarting app\n\n"
+  python3 aideas/main.py  # --agents translation
+fi