+++ shell/run.web.sh
@@ -1,7 +1,0 @@
-#!/usr/bin/env bash
-
-source ./pre_run.sh
-
-printf "\nStarting web app\n\n"
-
-python3 aideas/web.py