+++ src/aideas/app/config_loader.py
@@ -15,23 +15,21 @@ _SUFFIX = '.config'
 
 
 class ConfigLoader(YamlLoader):
-    def __init__(self, config_path: str, run_config: dict[str, any] = None):
+    def __init__(self, config_path: str, variable_source: dict[str, any] = Env.collect()):
         super().__init__(config_path, suffix=_SUFFIX)
-        self.__variable_source = {}
-        self.__variable_source.update(Env.collect())  # Environment variables
-        if run_config is None:
-            self.__variable_source.update(self.load_run_config())  # Properties file
-        self.__variable_source.update(RunArg.of_sys_argv())  # sys.argv
-        if run_config is not None:
-            self.__variable_source.update(run_config)  # User supplied
-        self.__agent_configs = self.load_agent_configs()
+        self.__config_path = config_path
+        self.__variable_source = variable_source
+        self.__agent_configs_with_un_replaced_variables = self.load_agent_configs(False)
+
+    def with_added_variable_source(self, source: dict[str, any]) -> 'ConfigLoader':
+        return ConfigLoader(self.__config_path, self.__variable_source)._add_variable_source(source)
 
     def get_sorted_agent_names(self,
                                config_filter: Callable[[dict[str, any]], bool],
                                config_sort: Callable[[dict[str, any]], int]) -> [str]:
         keys = []
         values = []
-        for k, v in self.__agent_configs.items():
+        for k, v in self.__agent_configs_with_un_replaced_variables.items():
             if config_filter(v):
                 keys.append(k)
                 values.append(v)
@@ -40,14 +38,36 @@ class ConfigLoader(YamlLoader):
 
         return [keys[values.index(sorted_val)] for sorted_val in values_sorted]
 
-    def load_agent_configs(self, cfg_filter=None) -> dict[str, dict[str, any]]:
+    def load_agent_configs(
+            self, check_replaced: bool = True, cfg_filter=None) -> dict[str, dict[str, any]]:
         configs = {}
         for name in self.__all_agent_names():
-            config = self.load_agent_config(name)
+            config = self.load_agent_config(name, check_replaced)
             if not cfg_filter or cfg_filter(config):
                 configs[name] = config
         logger.debug(f"Config names: {configs.keys()}")
         return configs
+
+    def load_run_config(self) -> dict[str, any]:
+        result = self.load_config("run")
+        return RunArg.of_dict(result)
+
+    def load_from_path(self, path: str, check_replaced: bool = True) -> dict[str, any]:
+        try:
+            return replace_all_variables(load_yaml(path), self.__variable_source, check_replaced)
+        except FileNotFoundError:
+            logger.warning(f'Could not find config file for: {path}')
+            return {}
+
+    def load_agent_config(self, agent_name: str, check_replaced: bool = True) -> dict[str, any]:
+        return self.load_from_path(self.get_agent_config_path(agent_name), check_replaced)
+
+    def get_agent_config_path(self, agent_name: str) -> str:
+        return self.get_path(os.path.join('agent', agent_name))
+
+    def _add_variable_source(self, source: dict[str, any]) -> 'ConfigLoader':
+        self.__variable_source.update(source)
+        return self
 
     def __all_agent_names(self) -> [str]:
         agents = []
@@ -56,19 +76,8 @@ class ConfigLoader(YamlLoader):
             agents.append(agent_filename[0:agent_filename.index(_SUFFIX)])
         return agents
 
-    def load_run_config(self) -> dict[str, any]:
-        result = self.load_config("run")
-        return RunArg.of_dict(result)
 
-    def load_from_path(self, path: str) -> dict[str, any]:
-        try:
-            return replace_all_variables(load_yaml(path), self.__variable_source)
-        except FileNotFoundError:
-            logger.warning(f'Could not find config file for: {path}')
-            return {}
-
-    def load_agent_config(self, agent_name: str) -> dict[str, any]:
-        return self.load_from_path(self.get_agent_config_path(agent_name))
-
-    def get_agent_config_path(self, agent_name: str) -> str:
-        return self.get_path(os.path.join('agent', agent_name))
+class SimpleConfigLoader(ConfigLoader):
+    def __init__(self, config_path: str):
+        super().__init__(config_path)
+        super()._add_variable_source(RunArg.of_sys_argv())  # sys.argv