+++ src/test/app/test_functions.py
@@ -13,12 +13,19 @@ from aideas.app.web.webdriver_creator import WebDriverCreator
 __TEST_SRC_DIR = f'{os.getcwd()}/test/app'
 
 
+class TestConfigLoader(ConfigLoader):
+    def __init__(self, config_path: str):
+        super().__init__(config_path)
+        super()._add_variable_source(self.load_run_config())  # Properties file
+        super()._add_variable_source(RunArg.of_sys_argv())  # sys.argv
+
+
 def get_main_config_loader() -> ConfigLoader:
-    return ConfigLoader(os.path.join("resources", "config"))
+    return TestConfigLoader(os.path.join("resources", "config"))
 
 
 def get_test_config_loader() -> ConfigLoader:
-    return ConfigLoader(os.path.join("test", "resources", "config"))
+    return TestConfigLoader(os.path.join("test", "resources", "config"))
 
 
 def load_agent_names() -> [str]: